+++ TPQUAD/source/Biquad/Biquad.h
@@ -9,13 +9,13 @@
 #define BIQUAD_BIQUAD_H_
 
 typedef struct {
-    double b0, b1, b2, a1, a2; // filter coefficients
-    double x1, x2, y1, y2;     // filter state variables
+	float b0, b1, b2, a1, a2; // filter coefficients
+	float x1, x2, y1, y2;     // filter state variables
 } BiQuad;
 
-void BiQuad_init(BiQuad* bq, double b[3], double a[3]);
+void BiQuad_init(BiQuad* bq, float b[3], float a[3]);
 
-double BiQuad_filter(BiQuad* bq, double input);
+float BiQuad_filter(BiQuad* bq, float input);
 
 
 #endif /* BIQUAD_BIQUAD_H_ */
