+++ TPQUAD/source/Sistema de Control/MatrixOps.c
@@ -8,11 +8,11 @@
 #include "MatrixOps.h"
 
 
-void matrix_mult(uint8_t m, uint8_t n, uint8_t p, const double A[m][n], const double B[n][p], double C[m][p])
+void matrix_mult(uint8_t m, uint8_t n, uint8_t p, const float A[m][n], const float B[n][p], float C[m][p])
 {
 	for (uint8_t i = 0; i < m; i++) {
 		for (uint8_t j = 0; j < p; j++) {
-			double sum = 0.0;
+			float sum = 0.0;
 			for (uint8_t k = 0; k < n; k++) {
 				sum += A[i][k] * B[k][j];
 			}
@@ -21,7 +21,7 @@ void matrix_mult(uint8_t m, uint8_t n, uint8_t p, const double A[m][n], const do
 	}
 }
 
-void scalar_mult(uint8_t m, uint8_t n, double scalar, double InOut[m][n]){
+void scalar_mult(uint8_t m, uint8_t n, float scalar, float InOut[m][n]){
 	for(uint8_t i = 0; i < m; i++){
 		for(uint8_t j = 0; j < n; j++){
 			InOut[i][j] = scalar*InOut[i][j];
@@ -29,7 +29,7 @@ void scalar_mult(uint8_t m, uint8_t n, double scalar, double InOut[m][n]){
 	}
 }
 
-void matrix_add_sub(uint8_t m, uint8_t n, const double A[m][n], char sign, const double B[m][n], double C[m][n]) {
+void matrix_add_sub(uint8_t m, uint8_t n, const float A[m][n], char sign, const float B[m][n], float C[m][n]) {
 	int8_t signtmp = sign == '-'? -1 : 1;
     for (uint8_t i = 0; i < m; i++) {
         for (uint8_t j = 0; j < n; j++) {
